Trasformations refactoring - New Directory
