+++ project/delivery_1/team4/README.md
@@ -72,8 +72,8 @@ graph
     RTS[Real-Time Server]
 
     subgraph Microservices
-		US[User service]
 		MS[Matchmaking service]
+		US[User service]
 	end
 
     GS[Game engine service]
@@ -84,10 +84,10 @@ graph
     CL --> GW
     CL --> RTS
     GW --> US
+    RTS --> US
     GW --> MS
 
     MS --> Queue1
-    GS --> Queue1
 
     RTS --> Broker1
     GS --> Broker1
