+++ project/delivery_1/team4/transformations/database.py
@@ -1,0 +1,4 @@
+from transformations.templates import generate_templated_component
+
+def generate_database(name):
+    generate_templated_component(name, "database")
