+++ project/delivery_1/team4/transformations/dockercompose.py
@@ -1,0 +1,4 @@
+from transformations.templates import _make_component_dir
+
+def generate_docker_compose(components):
+    path = _make_component_dir("")
